requirements.txt
flask
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
faiss-cpu
lightgbm
torch --index-url https://download.pytorch.org/whl/cpu
transformers==4.30.2
sentence-transformers==2.2.2
